+++ tests/integration/normal_mode_test.py
@@ -535,6 +535,17 @@ def test_e_key(r: TmuxRunner):
 
 
 @setup("tests/fixture/test_file_1.txt")
+def test_left_chevron_key(r: TmuxRunner):
+    r.press('<')
+    assert "[X]" not in r.statusbar_parts()
+
+    r.press('j')
+    r.press('<')
+    assert "[X]" in r.statusbar_parts()
+    assert "2\u2502here" in r.lines()[1]
+
+
+@setup("tests/fixture/test_file_1.txt")
 def test_right_chevron_key(r: TmuxRunner):
     r.press('>')
     assert "    This" in r.lines()[0]