+++ tests/model_test.cpp
@@ -339,15 +339,18 @@ TEST_CASE("find_prev", "[model]") {
 
     auto ret = m.find_prev('f');
     REQUIRE(ret.has_value());
-    REQUIRE(ret.value().vertical == 0);
+    REQUIRE(ret.value().vertical == 5);
     REQUIRE(ret.value().horizontal == 5);
 
-    m.current_line -= static_cast<unsigned int>(ret.value().vertical);
-    m.current_char -= static_cast<unsigned int>(ret.value().horizontal);
+    m.current_line = static_cast<unsigned int>(ret.value().vertical);
+    m.current_char = static_cast<unsigned int>(ret.value().horizontal);
+    REQUIRE(m.current_line == 5);
+    REQUIRE(m.current_char == 5);
+    REQUIRE(m.buf.at(5).at(5) == 'f');
 
     ret = m.find_prev('t');
     REQUIRE(ret.has_value());
-    REQUIRE(ret.value().vertical == 3);
+    REQUIRE(ret.value().vertical == 2);
     REQUIRE(ret.value().horizontal == 5);
 
     m.current_line -= static_cast<unsigned int>(ret.value().vertical);
@@ -712,4 +715,20 @@ TEST_CASE("indent_curr_line", "[model]") {
 
     REQUIRE(m.buf.at(0).at(0) == ' ');
     REQUIRE(m.buf.at(0).at(4) == 'f');
+}
+
+TEST_CASE("dedent_curr_line", "[model]") {
+    auto m = Model({"    foo", "bar"}, "");
+
+    // Dedent a line that has some indentation
+    m.dedent_curr_line();
+    REQUIRE(m.buf.at(0).size() == 3);
+    REQUIRE(m.buf.at(1).size() == 3);
+    REQUIRE(m.buf.at(0).at(0) == 'f');
+
+    // Dedent a line that has no indentation
+    m.dedent_curr_line();
+    REQUIRE(m.buf.at(0).size() == 3);
+    REQUIRE(m.buf.at(1).size() == 3);
+    REQUIRE(m.buf.at(0).at(0) == 'f');
 }